execute drainer also on delete events
