+++ service/controller/resource/drainer/delete.go
@@ -2,10 +2,15 @@ package drainer
 
 import (
 	"context"
+
+	"github.com/giantswarm/microerror"
 )
 
-// EnsureDeleted is a no-op, because the lifecycle resource only has to act on
-// create and update events in order to drain guest cluster nodes.
 func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	err := r.ensure(ctx, obj)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
 	return nil
 }
